73rd commit

+++ recipe/views.py
@@ -27,9 +27,22 @@ def recipe_detail_view(request,id=None):
 @login_required
 def recipe_delete_view(request,id=None):
     obj=get_object_or_404(Recipe,id=id)
+    try:
+        obj = get_object_or_404(Recipe, id=id, )
+    except:
+        obj = None
+    if obj is None:
+        if request.htmx:
+            return HttpResponse("Not Found")
+        raise Http404
     if request.method == 'POST':
         obj.delete()
         success_url=reverse("recipe:list")
+        if request.htmx:
+            header = {
+                'HX-Redirect': success_url
+            }
+            return HttpResponse("Success", headers=header)
         return redirect(success_url)
     context={
         "object":obj
@@ -39,10 +52,20 @@ def recipe_delete_view(request,id=None):
 
 @login_required
 def recipe_ingredient_delete_view(request,parent_id=None,id=None):
-    obj=get_object_or_404(RecipeIngredients, parent__id=parent_id,id=id)
+    try:
+        obj = RecipeIngredients.objects.get(recipe__id=parent_id, id=id, recipe__user=request.user)
+    except:
+        obj = None
+    if obj is None:
+        if request.htmx:
+            return HttpResponse("Not Found")
+        raise Http404
     if request.method == 'POST':
+        name=obj.name
         obj.delete()
         success_url=reverse("recipe:detail",kwargs={"id":parent_id})
+        if request.htmx:
+            return render(request, "recipes/partials/ingredient-inline-delete-response.html", {"name": name})
         return redirect(success_url)
     context={
         "object":obj
